+++ src/gui/groupchatwindow.c
@@ -579,7 +579,7 @@ static void qq_group_chatwindow_init(QQGroupChatWindow *win)
 static void qq_group_chatwindow_getter(GObject *object, guint property_id,  
                                     GValue *value, GParamSpec *pspec)
 {
-    if(object == NULL || value == NULL || property_id < 0){
+    if(object == NULL || value == NULL || (gint)property_id < 0){
             return;
     }
     
@@ -604,7 +604,7 @@ static void qq_group_chatwindow_getter(GObject *object, guint property_id,
 static void qq_group_chatwindow_setter(GObject *object, guint property_id,  
                                  const GValue *value, GParamSpec *pspec)
 {
-    if(object == NULL || value == NULL || property_id < 0){
+    if(object == NULL || value == NULL || (gint)property_id < 0){
             return;
     }
     QQGroupChatWindowPriv *priv = G_TYPE_INSTANCE_GET_PRIVATE(
